:some tweaking and reducing of interval needed, also routes for scheudler seem to return error even in valid scenario, suspicious

+++ lib/models/LeanStat.h
@@ -7,5 +7,5 @@ struct LeanStat
 {
     String capturedAt;
 
-    int leanAngle;
+    float leanAngle;
 };
